+++ packages/pocket-ic/CHANGELOG.md
@@ -14,7 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The library functions `PocketIc::install_canister`, `PocketIc::upgrade_canister`, and `PocketIc::reinstall_canister`
   support installing canisters with a large WASM as a sequence of chunks (transparently, i.e.,
   the user does not need to take any extra action).
-- The maximum duration (timeout) of a PocketIC operation is configurable and can be deactivated by specifying it as `None` (the default is a timeout of 5 minutes).
+- The maximum duration (timeout) of a PocketIC operation is configurable using `PocketIcBuilder::with_max_request_time_ms`
+  and can be deactivated by specifying it as `None` (the default is a timeout of 5 minutes).
 - The library function `PocketIc::create_canister_with_id` works for all IC mainnet canister IDs that do not belong to the NNS or II subnet.
 - The library function `PocketIc::uninstall_canister` to uninstall code of an existing canister.
 - The library function `PocketIc::update_canister_settings` to update settings (e.g., compute allocation) of an existing canister.
@@ -22,8 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   an arbitrary unassigned port) and optionally specified domains (default to `localhost`) and using an optionally specified TLS certificate (if provided,
   an HTTPS gateway is created) and configures the PocketIC instance to make progress automatically, i.e., periodically update the time of the PocketIC instance to the real time
   and execute rounds on the subnets.
-- The library function `PocketIc::make_live_https` configuring a PocketIc instance to automatically make progress (updating time and executing rounds)
-  and creating an HTTPS gateway for that instance listening at a dedicated domain and port and using a specified TLS certificate.
 - The function `PocketIcBuilder::with_server_url` to specify the URL of the PocketIC server (if not used, then the URL of an already running PocketIC server
   is derived or a new PocketIC server is started).
 - The function `PocketIcBuilder::with_state_dir` to specify a directory in which the state of the PocketIC instance can be preserved across the PocketIC instance lifetime
@@ -35,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Deprecated `make_deterministic`, use `stop_live` instead
+- The topology contains a list of node IDs instead of subnet size which can be derived from the number of node IDs.